requirements.txt
streamlit==1.36.0
pandas==2.0.3
openpyxl==3.10.10
openai==2.26.0
python-dotenv==1.0.0
PyPDF2==4.3.1
python-pptx==0.6.23
tqdm==4.66.1